requirements.txt
streamlit
groq
openai
pinecone-client
python-dotenv
